+++ src/main/java/org/jenkinsci/plugins/ghprb/jobdsl/GhprbTriggerContext.java
@@ -42,6 +42,8 @@ class GhprbTriggerContext implements Context {
 
     boolean displayBuildErrorsOnDownstreamBuilds;
 
+    String commentFilePath;
+
     String buildDescriptionTemplate;
 
     String includedRegions;
@@ -275,6 +277,20 @@ class GhprbTriggerContext implements Context {
      */
     public void displayBuildErrorsOnDownstreamBuilds() {
         displayBuildErrorsOnDownstreamBuilds(true);
+    }
+
+    /**
+     * Pubilsh this comment file to the github PR
+     */
+    public void commentFilePath(String commentFilePath) {
+        this.commentFilePath = commentFilePath;
+    }
+
+    /**
+     * Pubilsh this comment file to the github PR
+     */
+    public void commentFilePath() {
+        commentFilePath("");
     }
 
     /**